feat: add info to help recover commit message
When husky fails, it's easy to think that the message you worked so hard on has been lost forever.
Luckily, that's usually not the case! When `git commit` opens your editor, it opens it pointing at
`.git/COMMIT_EDITMSG`. All you need to do to recover your message is to take a look at that file.
To prevent this from being a problem going forward, we can just include a message on failure.